Configure first example tab name
Added an exampleOpts prop. Consider refactoring to move all example
config into this object and rename. In the same vein, it would be nice
to be able to configure the style guide via pure JSON until you can’t
at which point, you’d revert to the current component with render
function style.